Socket and Stdio transports are Backport adapters.
