+++ lib/solargraph/language_server/transport.rb
@@ -7,7 +7,6 @@ module Solargraph
       autoload :DataReader, 'solargraph/language_server/transport/data_reader'
       autoload :Socket,     'solargraph/language_server/transport/socket'
       autoload :Stdio,      'solargraph/language_server/transport/stdio'
-      autoload :BackportTcp, 'solargraph/language_server/transport/backport_tcp'
     end
   end
 end
